# carsearch/src/main/resources/api/api.yaml
openapi: 3.1.0
servers:
    - url: http://localhost:8080
      description: Local server
info:
    title: CarDB API
    version: 0.0.1
    description: API for searching car models, variations, engine options, and specs

paths:
    /cars:
        get:
            summary: Get cars
            description: Retrieves a list of car models
            operationId: getCars
            parameters:
                - in: query
                  name: search
                  schema:
                      type: string
                  description: The search string
                - in: query
                  name: brandIds
                  schema:
                      type: array
                      items:
                        type: string
                        format: uuid
                  description: The ids of the brands to filter by
                - in: query
                  name: body_types
                  schema:
                      type: array
                      items:
                          $ref: '#/components/schemas/ApiBodyType'
                  description: The body types to filter by
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ApiCarModel'
                            examples:
                                default_example:
                                    value:
                                        cars:
                                            - id: 1
                                              name: 'Audi A4'
                                              slug: 'audi-a4'
                                              body_type: 'SEDAN'
                                              brand:
                                                id: 1
                                                name: 'Audi'
                                                slug: 'audi'
                                                origin_country: 'Germany'
                                                website: 'https://www.audi.com'
                                            - id: 2
                                              name: 'BMW 3 Series'
                                              slug: 'bmw-3-series'
                                              body_type: 'SEDAN'
                                              brand:
                                                id: 2
                                                name: 'BMW'
                                                slug: 'bmw'
                                                origin_country: 'Germany'
                                                website: 'https://www.bmw.com'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 401
                                        message: 'Unauthorized'
        post:
            summary: Create car
            description: Creates a new car model
            operationId: createCar
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiCreateCarModel'
            responses:
                '201':
                    description: Car created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiCarModel'
                            examples:
                                default_example:
                                    value:
                                        id: 1
                                        name: 'Audi A4'
                                        slug: 'audi-a4'
                                        body_type: 'SEDAN'
                                        brand:
                                            id: 1
                                            name: 'Audi'
                                            slug: 'audi'
                                            origin_country: 'Germany'
                                            website: 'https://www.audi.com'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 401
                                        message: 'Unauthorized'

    /cars/{slug}:
        get:
            summary: Get car
            description: Retrieves a single car model
            operationId: getCar
            parameters:
                - name: slug
                  in: path
                  description: Slug of car model to retrieve
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiCarModel'
                            examples:
                                skoda_octavia_combi:
                                    summary: Skoda Octavia Combi
                                    value:
                                        id: 1
                                        name: 'Skoda Octavia Combi'
                                        slug: 'skoda-octavia-combi'
                                        body_type: 'WAGON'
                                        brand:
                                            id: 1
                                            name: 'Skoda'
                                            slug: 'skoda'
                                            origin_country: 'Czech Republic'
                                            website: 'https://www.skoda-auto.com'
                                                

                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 401
                                        message: 'Unauthorized'
                '404':
                    description: Car model not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 404
                                        message: 'Car model not found'

    /cars/{slug}/generations:
        get:
            summary: Get car generations
            description: Retrieves a list of car generations
            operationId: getGenerations
            parameters:
                - name: slug
                  in: path
                  description: Slug of the car model to retrieve generations for
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ApiCarGeneration'
                            examples:
                                default_example:
                                    value:
                                        generations:
                                            - id: 1
                                              name: 'Skoda Octavia Combi 2019 - 2024'
                                              production_start_year: 2019
                                              production_end_year: 2024
                                              tire_size: '205/55 R16'
                                              length_mm: 4689
                                              width_mm: 1829
                                              height_mm: 1468
                                              ground_clearance_mm: 78
                                              cargo_volume_litres: 640

        post:
            summary: Create car generation
            description: Creates a new car generation
            operationId: createGeneration
            parameters:
                - name: slug
                  in: path
                  description: Slug of the car model to create a generation for
                  required: true
                  schema:
                      type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiCreateCarGeneration'
            responses:
                '201':
                    description: Generation created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiCarGeneration'
                            examples:
                                default_example:
                                    value:
                                        id: 1
                                        name: 'Skoda Octavia Combi 2019 - 2024'
                                        production_start_year: 2019
                                        production_end_year: 2024
                                        tire_size: '205/55 R16'
                                        length_mm: 4689
                                        width_mm: 1829
                                        height_mm: 1468
                                        ground_clearance_mm: 78
                                        cargo_volume_litres: 640

    /brands:
        get:
            summary: Get car brands
            description: Retrieves a list of car brands
            operationId: getBrands
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: '#/components/schemas/ApiCarBrand'
                            examples:
                                default_example:
                                    value:
                                        brands:
                                            - id: 1
                                              name: 'Audi'
                                              origin_country: 'Germany'
                                              website: 'https://www.audi.com'
                                            - id: 2
                                              name: 'BMW'
                                              origin_country: 'Germany'
                                              website: 'https://www.bmw.com'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
        post:
            summary: Create car brand
            description: Creates a new car brand
            operationId: createBrand
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/ApiCreateCarBrand'
            responses:
                '201':
                    description: Brand created
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiCarBrand'
                            examples:
                                default_example:
                                    value:
                                        id: 1
                                        name: 'Audi'
                                        origin_country: 'Germany'
                                        website: 'https://www.audi.com'
                                        slug: 'audi'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 401
                                        message: 'Unauthorized'
    /brands/{slug}:
        get:
            summary: Get car brand
            description: Retrieves a single car brand
            operationId: getBrand
            parameters:
                - name: slug
                  in: path
                  description: Slug of the car brand to retrieve
                  required: true
                  schema:
                      type: string
            responses:
                '200':
                    description: Successful response
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiCarBrand'
                            examples:
                                default_example:
                                    value:
                                        id: 1
                                        name: 'Audi'
                                        origin_country: 'Germany'
                                        website: 'https://www.audi.com'
                '400':
                    description: Bad request
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 400
                                        message: 'Bad request'
                '401':
                    description: Unauthorized
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 401
                                        message: 'Unauthorized'
                '404':
                    description: Car brand not found
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/ApiError'
                            examples:
                                default_example:
                                    value:
                                        statusCode: 404
                                        message: 'Car brand not found'
components:
    schemas:
        ApiCarModel:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                slug:
                    type: string
                body_type:
                    $ref: '#/components/schemas/ApiBodyType'
                brand:
                    $ref: '#/components/schemas/ApiCarBrand'

        ApiCreateCarModel:
            type: object
            properties:
                name:
                    type: string
                body_type:
                    $ref: '#/components/schemas/ApiBodyType'
                brand_id:
                    type: string
                    format: uuid
                slug:
                    type: string

        ApiCarGeneration:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                production_start_year:
                    type: integer
                production_end_year:
                    type:
                        - integer
                        - null
                tire_size:
                    type: string
                length_mm:
                    type: integer
                width_mm:
                    type: integer
                height_mm:
                    type: integer
                ground_clearance_mm:
                    type: integer
                cargo_volume_litres:
                    type: integer
                # Add weight

        ApiCreateCarGeneration:
            type: object
            properties:
                name:
                    type: string
                production_start_year:
                    type: integer
                production_end_year:
                    type: integer
                tire_size:
                    type: string
                length_mm:
                    type: integer
                width_mm:
                    type: integer
                height_mm:
                    type: integer
                ground_clearance_mm:
                    type: integer
                cargo_volume_litres:
                    type: integer
                # Add weight

        ApiEngineOption:
            type: object
            properties:
                id:
                    type: integer
                name:
                    type: string
                fuel_type:
                    $ref: '#/components/schemas/ApiFuelType'
                drive_type:
                    $ref: '#/components/schemas/ApiDriveType'
                transmission_type:
                    $ref: '#/components/schemas/ApiTransmissionType'
                number_of_gears:
                    type: integer
                number_of_cylinders:
                    type: integer
                cylinder_layout:
                    $ref: '#/components/schemas/ApiCylinderLayout'
                fuel_economy_city_l_per_100km:
                    type: number
                fuel_economy_highway_l_per_100km:
                    type: number
                fuel_economy_combined_l_per_100km:
                    type: number
                power_kw:
                    type: number
                power_hp:
                    type: number
                power_bhp:
                    type: number
                top_speed_kph:
                    type: integer
                acceleration_0_100_kph:
                    type: number

        ApiCarBrand:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                name:
                    type: string
                origin_country:
                    type: string
                website:
                    type: string
                slug:
                    type: string

        ApiCreateCarBrand:
            type: object
            properties:
                name:
                    type: string
                origin_country:
                    type: string
                website:
                    type: string
                slug:
                    type: string

        ApiBodyType:
            type: string
            enum:
                - SEDAN
                - COUPE
                - CONVERTIBLE
                - WAGON
                - HATCHBACK
                - SUV
                - TRUCK
                - VAN

        ApiFuelType:
            type: string
            enum:
                - PETROL
                - DIESEL
                - ELECTRIC
                - HYBRID
                - PLUGIN_HYBRID
                - LPG
                - CNG
                - HYDROGEN

        ApiDriveType:
            type: string
            enum:
                - FWD
                - RWD
                - AWD

        ApiTransmissionType:
            type: string
            enum:
                - MANUAL
                - AUTOMATIC
                - SEMI_AUTOMATIC
                - CVT
                - DCT

        ApiCylinderLayout:
            type: string
            enum:
                - INLINE
                - V
                - FLAT
                - ROTARY
                - ELECTRIC
                - HYDROGEN
                - HYBRID

        ApiError:
            type: object
            properties:
                statusCode:
                    type: integer
                message:
                    type: string